% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.cca.R
\docType{data}
\name{simData.cca}
\alias{simData.cca}
\title{Example data for method iscca}
\format{list}
\description{
Example data for users to apply the method iscca, iscca.cv, meta.scca or scca.
}
\keyword{datasets}
