% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.pca.R
\docType{data}
\name{simData.pca}
\alias{simData.pca}
\title{Example data for method ispca}
\format{list}
\description{
Example data for users to apply the method ispca, ispca.cv, meta.spca or spca.
}
\keyword{datasets}
