% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.pls.R
\docType{data}
\name{simData.pls}
\alias{simData.pls}
\title{Example data for method ispls}
\format{list}
\description{
Example data for users to apply the method ispls, ispls.cv, meta.spls or spls.
}
\keyword{datasets}
